Organize application into package
