+++ .gitignore
@@ -1,2 +1,5 @@
 /.venv
 .env
+__pycache__/
+*.py[cod]
+cotizacion.pdf
