+++ app/database.py
@@ -1,8 +1,7 @@
-from sqlalchemy import create_engine, Column, Integer, String
+from sqlalchemy import create_engine
 from sqlalchemy.engine import URL
 from sqlalchemy.orm import declarative_base, sessionmaker
 from decouple import config
-
 
 url = URL.create(
     drivername="postgresql",
@@ -10,20 +9,11 @@ url = URL.create(
     password=config("DB_PASSWORD"),
     host="localhost",
     database="mydb",
-    port=5432
+    port=5432,
 )
 
 engine = create_engine(url)
 SessionLocal = sessionmaker(bind=engine)
 Base = declarative_base()
-
-class Conversation(Base):
-    __tablename__ = "conversations"
-
-    id = Column(Integer, primary_key=True, index=True)
-    sender = Column(String)
-    message = Column(String)
-    response = Column(String)
-
 
 Base.metadata.create_all(engine)
