+++ app/main.py
@@ -1,6 +1,5 @@
 # Third-party imports
 import openai
-import os
 import time
 from fastapi import FastAPI, Form, Depends
 from fastapi.responses import FileResponse
@@ -9,37 +8,10 @@ from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
 # Internal imports
-from models import Conversation, SessionLocal
-from utils import send_message, logger
-
-from fpdf import FPDF
-
-current_directory = os.getcwd()
-print(current_directory)
-
-def generate_pdf(content):
-    class PDF(FPDF):
-        def header(self):
-            # Seleccionar fuente Arial bold 15
-            self.set_font('Arial', 'B', 15)
-            # Mover a la derecha
-            self.cell(80)
-            # Título
-            self.cell(30, 10, 'Cotizacion', 1, 0, 'C')
-            # Salto de línea
-            self.ln(20)
-
-    # Instancia de la clase para construir el PDF
-    pdf = PDF()
-    pdf.add_page()
-    pdf.set_font("Arial", size = 12)
-    pdf.multi_cell(0, 10, content)
-    
-    file_name = os.path.join("cotizacion.pdf") 
-
-    pdf.output(file_name)
-    return file_name
-
+from app.models import Conversation
+from app.database import SessionLocal
+from app.utils import send_message, logger
+from app.services.pdf import generate_pdf
 
 app = FastAPI()
 # Set up the OpenAI API client
@@ -62,7 +34,6 @@ async def reply(Body: str = Form(), db: Session = Depends(get_db)):
     elif "cotización" in Body.lower():
         content = "Detalle de Cotización:\nEl precio del martillo es de $3 dólares y hay 4 unidades disponibles."
         pdf_file_path = generate_pdf(content)
-        print(pdf_file_path)
         return FileResponse(pdf_file_path, media_type="application/pdf", filename="/cotizacion.pdf")
     else:
         chat_response = ""
@@ -93,14 +64,13 @@ async def reply(Body: str = Form(), db: Session = Depends(get_db)):
         else:
             chat_response = "Lo siento, no puedo responder en este momento."
 
-
     # Store the conversation in the database
     try:
         conversation = Conversation(
             sender=whatsapp_number,
             message=Body,
             response=chat_response
-            )
+        )
         db.add(conversation)
         db.commit()
         logger.info(f"Conversation #{conversation.id} stored in database")
