+++ app/models.py
@@ -1,0 +1,11 @@
+from sqlalchemy import Column, Integer, String
+from app.database import Base
+
+
+class Conversation(Base):
+    __tablename__ = "conversations"
+
+    id = Column(Integer, primary_key=True, index=True)
+    sender = Column(String)
+    message = Column(String)
+    response = Column(String)
